# console/console.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Nakama console API v2",
    "version": "2.0",
    "contact": {
      "name": "The Nakama Authors \u0026 Contributors",
      "url": "https://github.com/heroiclabs/nakama",
      "email": "hello@heroiclabs.com"
    }
  },
...
  "host": "127.0.0.1:7351",
  "schemes": [
    "http"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/v2/console/authenticate": {
      "post": {
        "summary": "Authenticate a console user with username and password.",
        "operationId": "Authenticate",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/ConsoleSession"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpc.Status"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "description": "Authenticate a console user with username and password.",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/AuthenticateRequest"
            }
          }
        ],
        "tags": [
          "Console"
        ],
        "security": [
          {
            "": []
          }
        ]
      }
    },
    "/v2/console/authenticate/logout": {
      "post": {
        "summary": "Log out a session and invalidate the session token.",
        "operationId": "AuthenticateLogout",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "type": "object",
              "properties": {}
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpc.Status"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "description": "Log out a session and invalidate a session token.",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/AuthenticateLogoutRequest"
            }
          }
        ],
        "tags": [
          "Console"
        ]
      }
    },
    "/v2/console/authenticate/mfa": {
      "post": {
        "summary": "Change an account's MFA using a code, usually delivered over email.",
        "operationId": "AuthenticateMFASetup",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/AuthenticateMFASetupResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpc.Status"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "description": "Request to change MFA.",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/AuthenticateMFASetupRequest"
            }
          }
        ],
        "tags": [
          "Console"
        ]
      }
    },
...
  },
  "definitions": {
...
    "Any": {
      "type": "object",
      "properties": {
        "@type": {
          "type": "string"
        }
      },
      "additionalProperties": {}
    },
...
    "AuthenticateLogoutRequest": {
      "type": "object",
      "properties": {
        "token": {
          "type": "string",
          "description": "Session token to log out."
        }
      },
      "description": "Log out a session and invalidate a session token."
    },
    "AuthenticateMFASetupRequest": {
      "type": "object",
      "properties": {
        "mfa": {
          "type": "string",
          "description": "MFA token."
        },
        "code": {
          "type": "string",
          "description": "MFA code."
        }
      },
      "description": "Request to change MFA."
    },
    "AuthenticateMFASetupResponse": {
      "type": "object",
      "properties": {
        "recovery_codes": {
          "type": "array",
          "items": {
            "type": "string"
          },
          "title": "An one-time code to configure the MFA mechanism"
        }
      },
      "description": "Response to change MFA."
    },
    "AuthenticateRequest": {
      "type": "object",
      "properties": {
        "username": {
          "type": "string",
          "description": "The username of the user."
        },
        "password": {
          "type": "string",
          "description": "The password of the user."
        },
        "mfa": {
          "type": "string",
          "description": "Multi-factor authentication code."
        }
      },
      "description": "Authenticate a console user with username and password."
    },
...
    "ConsoleSession": {
      "type": "object",
      "properties": {
        "token": {
          "type": "string",
          "description": "A session token (JWT) for the console user."
        },
        "mfa_code": {
          "type": "string",
          "description": "MFA code required to setup the MFA mechanism."
        }
      },
      "description": "A console user session."
    },
...
    "rpc.Status": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/Any"
          }
        }
      }
    }
  },
  "securityDefinitions": {
    "BasicAuth": {
      "type": "basic"
    },
    "BearerJwt": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header"
    }
  },
  "security": [
    {
      "BearerJwt": []
    }
  ],
...
}